python/sk_financial_analyst/deployment/durable_function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity>=1.16.0
python-dotenv>=0.10.3
azure-keyvault-secrets
pyyaml
requests
edgartools
yfinance==0.2.40
semantic-kernel
pydantic==2.8.2
